+++ x/foundation/client/testutil/query.go
@@ -11,6 +11,7 @@ import (
 	sdk "github.com/line/lbm-sdk/types"
 	"github.com/line/lbm-sdk/x/foundation"
 	"github.com/line/lbm-sdk/x/foundation/client/cli"
+	stakingtypes "github.com/line/lbm-sdk/x/staking/types"
 )
 
 func (s *IntegrationTestSuite) TestNewQueryCmdParams() {
@@ -29,13 +30,16 @@ func (s *IntegrationTestSuite) TestNewQueryCmdParams() {
 			[]string{},
 			true,
 			&foundation.QueryParamsResponse{
-				Params: &foundation.Params{
-					Enabled:       true,
+				Params: foundation.Params{
 					FoundationTax: sdk.MustNewDecFromStr("0.2"),
+					CensoredMsgTypeUrls: []string{
+						sdk.MsgTypeURL((*stakingtypes.MsgCreateValidator)(nil)),
+						sdk.MsgTypeURL((*foundation.MsgWithdrawFromTreasury)(nil)),
+					},
 				},
 			},
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				"extra",
 			},
@@ -78,7 +82,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdTreasury() {
 			[]string{},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				"extra",
 			},
@@ -119,7 +123,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdFoundationInfo() {
 			[]string{},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				"extra",
 			},
@@ -159,16 +163,17 @@ func (s *IntegrationTestSuite) TestNewQueryCmdMember() {
 	}{
 		"valid query": {
 			[]string{
-				val.Address.String(),
+				s.permanentMember.String(),
 			},
 			true,
 			&foundation.Member{
-				Address:  val.Address.String(),
-				Metadata: "genesis member",
+				Address:  s.permanentMember.String(),
+				Metadata: "permanent member",
 			},
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
+				s.permanentMember.String(),
 				"extra",
 			},
 			false,
@@ -210,7 +215,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdMembers() {
 			[]string{},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				"extra",
 			},
@@ -249,19 +254,15 @@ func (s *IntegrationTestSuite) TestNewQueryCmdProposal() {
 	}{
 		"valid query": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 			},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 				"extra",
 			},
-			false,
-		},
-		"not enough args": {
-			[]string{},
 			false,
 		},
 	}
@@ -299,7 +300,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdProposals() {
 			[]string{},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				"extra",
 			},
@@ -338,22 +339,16 @@ func (s *IntegrationTestSuite) TestNewQueryCmdVote() {
 	}{
 		"valid query": {
 			[]string{
-				"1",
-				val.Address.String(),
+				fmt.Sprintf("%d", s.proposalID),
+				s.permanentMember.String(),
 			},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
-				"1",
-				val.Address.String(),
+				fmt.Sprintf("%d", s.proposalID),
+				s.permanentMember.String(),
 				"extra",
-			},
-			false,
-		},
-		"not enough args": {
-			[]string{
-				"1",
 			},
 			false,
 		},
@@ -390,19 +385,15 @@ func (s *IntegrationTestSuite) TestNewQueryCmdVotes() {
 	}{
 		"valid query": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 			},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 				"extra",
 			},
-			false,
-		},
-		"not enough args": {
-			[]string{},
 			false,
 		},
 	}
@@ -438,19 +429,15 @@ func (s *IntegrationTestSuite) TestNewQueryCmdTallyResult() {
 	}{
 		"valid query": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 			},
 			true,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
-				"1",
+				fmt.Sprintf("%d", s.proposalID),
 				"extra",
 			},
-			false,
-		},
-		"not enough args": {
-			[]string{},
 			false,
 		},
 	}
@@ -500,7 +487,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdGrants() {
 			true,
 			1,
 		},
-		"extra args": {
+		"wrong number of args": {
 			[]string{
 				s.stranger.String(),
 				foundation.ReceiveFromTreasuryAuthorization{}.MsgTypeURL(),